change ImageProcessor! rename Image for Catalog Seed

+++ FlowersShopMVCTreaning/Controllers/ApiControllers/AdminController.cs
@@ -1,38 +1,52 @@
 ﻿using FlowersShopMVCTraining.Repository.Model;
 using FlowersShopMVCTraining.Repository.Repository;
 using FlowersShopMVCTraining.Service;
+    using Microsoft.AspNetCore.Http.HttpResults;
 using Microsoft.AspNetCore.Mvc;
 
 namespace FlowersShopMVCTraining.Controllers.ApiControllers
 {
     [ApiController]
     [Route("/api/[controller]/[action]")]
-    public class AdminController
+    public class AdminController : ControllerBase
     {
         private const string EXTENSION_IMAGE = ".jpg";
         private ProductDescriptionRepository _productDescriptionRepository;
         private ShopCardRepository _shopCardRepository;
         private PathHelper _pathHelper;
+        private ILogger<AdminController> _logger;
 
         public AdminController(ProductDescriptionRepository productDescriptionRepository, 
                                ShopCardRepository shopCardRepository,
-                               PathHelper pathHelper)
+                               PathHelper pathHelper,
+                               ILogger<AdminController> logger)
         {
             _productDescriptionRepository = productDescriptionRepository;
             _shopCardRepository = shopCardRepository;
             _pathHelper = pathHelper;
+            _logger = logger;
         }
-        public void DeleteCard(int cardId)
+        public IActionResult DeleteCard(int cardId)
         {
-            var imageName = _shopCardRepository.Get(cardId)!.ImageName + EXTENSION_IMAGE;
+            var shopCard = _shopCardRepository.Get(cardId);
+
+            if(shopCard == null)
+            {
+                _logger.LogError($"Удаление завершилось с ошибкой. Запрошенный в базе не существует {cardId}");
+                return BadRequest();
+            }
+
+            var imageName = shopCard.ImageName + EXTENSION_IMAGE;           
             var pathToSmall = _pathHelper.GetPathByFolder("img\\output\\small", imageName);
             var pathToLarge = _pathHelper.GetPathByFolder("img\\output\\large", imageName);
             var productDescriptionId = _shopCardRepository.GetDescriptionId(cardId);
 
             _productDescriptionRepository.Delete(productDescriptionId);
             _shopCardRepository.Delete(cardId);
-            File.Delete(pathToSmall);
-            File.Delete(pathToLarge);
+            System.IO.File.Delete(pathToSmall);
+            System.IO.File.Delete(pathToLarge);
+
+            return Ok();
         }
     }
 }
